Add files via upload
This Program accesses the camera and in this case default computer camera( If you have an external camera,
 use "video = cv2.VideoCapture(1)" in line 6.
The program starts and checks for QR or barcodes and keeps running itself until there is a QR Code or a Barcode. If the screen consists of 2 or more Codes, it will display all the QR details.